even more warnings patches from Michael D. Adams
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1543 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/CSeries/csmisc.cpp
@@ -35,7 +35,7 @@ Jan 25, 2002 (Br'fin (Jeremy Parsons)):
 
 extern void initialize_debugger(bool);
 
-unsigned long machine_tick_count(void)
+uint32 machine_tick_count(void)
 {
 	return TickCount();
 }
