+++ Source_Files/GameWorld/effects.h
@@ -173,10 +173,10 @@ effect_data *get_effect_data(
 // LP: to pack and unpack this data;
 // these do not make the definitions visible to the outside world
 
-uint8 *unpack_effect_data(uint8 *Stream, effect_data *Objects, int Count);
-uint8 *pack_effect_data(uint8 *Stream, effect_data *Objects, int Count);
-uint8 *unpack_effect_definition(uint8 *Stream, int Count);
-uint8 *pack_effect_definition(uint8 *Stream, int Count);
+uint8 *unpack_effect_data(uint8 *Stream, effect_data *Objects, size_t Count);
+uint8 *pack_effect_data(uint8 *Stream, effect_data *Objects, size_t Count);
+uint8 *unpack_effect_definition(uint8 *Stream, size_t Count);
+uint8 *pack_effect_definition(uint8 *Stream, size_t Count);
 
 #endif
